changed to floating point precision for movement

+++ src/Components/ColliderComponent.cpp
@@ -12,10 +12,10 @@ void ColliderComponent::initialize() {
 #endif  // DEBUG
     _transform = _entity->getComponent<TransformComponent>();
     if (_transform != nullptr) {
-        _collider.x = static_cast<int>(_transform->_position.x);
-        _collider.y = static_cast<int>(_transform->_position.y);
-        _collider.w = static_cast<int>(_transform->_width * _transform->_scale);
-        _collider.h = static_cast<int>(_transform->_height * _transform->_scale);
+        _collider.x = _transform->_position.x;
+        _collider.y = _transform->_position.y;
+        _collider.w = _transform->_width * _transform->_scale;
+        _collider.h = _transform->_height * _transform->_scale;
         _source = {0, 0, _transform->_width, _transform->_height};
         _destination = {_collider.x, _collider.y, _collider.w, _collider.h};
     }
@@ -23,10 +23,10 @@ void ColliderComponent::initialize() {
 
 void ColliderComponent::update(float delta_time) {
     if (_transform != nullptr) {
-        _collider.x = static_cast<int>(_transform->_position.x);
-        _collider.y = static_cast<int>(_transform->_position.y);
-        _collider.w = static_cast<int>(_transform->_width * _transform->_scale);
-        _collider.h = static_cast<int>(_transform->_height * _transform->_scale);
+        _collider.x = _transform->_position.x;
+        _collider.y = _transform->_position.y;
+        _collider.w = _transform->_width * _transform->_scale;
+        _collider.h = _transform->_height * _transform->_scale;
 
         _destination.x = _collider.x - Engine::_camera.x;
         _destination.y = _collider.y - Engine::_camera.y;
