+++ src/Components/TransformComponent.cpp
@@ -1,7 +1,7 @@
 #include "TransformComponent.hpp"
 
 // initialize our members
-TransformComponent::TransformComponent(float position_x, float position_y, float velocity_x, float velocity_y, float width, float height, float scale)
+TransformComponent::TransformComponent(float position_x, float position_y, float velocity_x, float velocity_y, int width, int height, int scale)
     : _position(glm::vec2(position_x, position_y)), _velocity(glm::vec2(velocity_x, velocity_y)), _width(width), _height(height), _scale(scale) {}
 
 // changes our position relevant to our time and velocity
